+++ plugins/action/network_device_images_info.py
@@ -34,8 +34,8 @@ argument_spec.update(
         networkDeviceUpdateStatus=dict(type="str"),
         sortBy=dict(type="str"),
         order=dict(type="str"),
-        offset=dict(type="float"),
-        limit=dict(type="float"),
+        offset=dict(type="int"),
+        limit=dict(type="int"),
         headers=dict(type="dict"),
     )
 )
